// src/enums/errorTypes.ts
export enum ErrorType {
  SERVICE_UNAVAILABLE = "SERVICE_UNAVAILABLE",
  DATABASE_ERROR = "DATABASE_ERROR",
  QUEUE_ERROR = "QUEUE_ERROR",
  VALIDATION_ERROR = "VALIDATION_ERROR",
  FILE_PROCESSING_ERROR = "FILE_PROCESSING_ERROR",
  UNKNOWN_ERROR = "UNKNOWN_ERROR",
  NOT_FOUND = "NOT_FOUND",
  UNAUTHORIZED = "UNAUTHORIZED",
  FORMAT_ERROR = "FORMAT_ERROR",
  FORBIDDEN = "FORBIDDEN",
}